refactor: extract event and criteria processing modules
